feat: refactor text and container components
