+++ src/utils/web/builder/registerComponents.ts
@@ -1,3 +1,3 @@
-import '@/components/ui/builder/text.builder'
-import '@/components/ui/builder/container.builder'
+import '@/components/ui/richText/text.builder'
+import '@/components/ui/container/container.builder'
 import '@/components/ui/button/button.builder'
